Update lib;
This patch adds a configurable defaultMessage to the opts object. The
defaultMessage is now happening outside of the ignoreUrls configuration
in order to make sure that a defaultMessage is set once per instance of
the library. The option to customize individual URLs is still available
by making an index of the ignoreUrls array an object and passing
configuration options in here.